pyproject.toml
[tool.poetry]
name = "bisnesmedia-integration-hub"
version = "0.1.0"
description = "Integration hub for 1C and MoySklad automation"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.31"}
alembic = "^1.13.1"
pydantic-settings = "^2.3.4"
httpx = "^0.27.0"
tenacity = "^8.4.1"
apscheduler = "^3.10.4"
streamlit = "^1.36.0"
python-telegram-bot = "^21.1.1"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-mock = "^3.14.0"
testcontainers = "^4.7.1"
ruff = "^0.5.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
